pyproject.toml
[tool.semantic_release]
upload_to_pypi = false
upload_to_release = false
version_toml = [
    "pyproject.toml:project.version",
]

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"
environment = { keep_trailing_newline = true }
exclude_commit_patterns = [
]
template_dir = "template"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "major",
    "feat",
    "fix",
    "chore",
    ":nut_and_bolt:",
    ":sparkles:",
]
major_tags = [
    "major",
]
minor_tags = [
    "feat",
    ":sparkles:",
]
patch_tags = [
    "fix",
    ":nut_and_bolt:",
]

[project]
dependencies = [
    "alembic>=1.18.3",
    "fastapi-versionizer>=4.0.2",
    "fastapi[standard]>=0.128.1",
    "greenlet>=3.3.1",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.3.2",
    "sqlmodel>=0.0.32",
]
description = "py-sync-it-api"
dynamic = [
    "version",
]
name = "py-sync-it-api"
readme = "README.md"
requires-python = ">=3.12" # Managed by the template

# TODO: add only if CLI
#[project.scripts]
#you-name-it = "you_name_it:main"

[dependency-groups]
dev = [
    "poethepoet",
    "pylint",
    "pylint-pydantic",
    "pyright",
    "pytest",
    "pytest-coverage",
    "pytest-mock",
    "pytest-unordered",
    "ruff",
    "taplo",
]

[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
    "versioningit>=3.1.2",
]

[tool.hatch.build.hooks.versioningit-onbuild]
build-file = "py_sync_it_api/__init__.py"
source-file = "src/py_sync_it_api/__init__.py"

[tool.hatch.build.targets.wheel]
packages = [
    "src/py_sync_it_api",
]

[tool.hatch.version]
source = "versioningit"

[tool.pyright]
reportMissingTypeArgument = true # Report generic classes used without type arguments
strictListInference = true       # Use union types when inferring types of lists elements, instead of Any
typeCheckingMode = "standard"

[tool.versioningit]
default-version = "0.0.0.dev0+0"

[tool.versioningit.vcs]
default-tag = "v0.0.0"

[tool.versioningit.format]
# Format used when there have been commits since the most recent tag:
distance = "{next_version}.dev{distance}+{vcs}{rev}"
# Example formatted version: 1.2.3.post42+ge174a1f

# Format used when there are uncommitted changes:
dirty = "{base_version}+dirty{build_date:%Y%m%d}"
# Example formatted version: 1.2.3+d20230922

# Format used when there are both commits and uncommitted changes:
distance-dirty = "{next_version}.dev{distance}+{vcs}{rev}.d{build_date:%Y%m%d}"
